requirements.txt
# Core Framework
streamlit==1.31.0

# AWS & Data
boto3==1.34.34
botocore==1.34.34
pandas==2.2.0
pyarrow==15.0.0

# Azure OpenAI
openai==1.12.0

# LangChain & LangGraph
langgraph==0.0.25
langchain==0.1.6
langchain-community==0.0.20
langchain-openai==0.0.5

# Vector Store
faiss-cpu==1.7.4

# Geospatial
geopandas==0.14.3
folium==0.15.1
shapely==2.0.3

# Schema & Validation
pydantic==2.6.1

# Utilities
python-dotenv==1.0.1
tiktoken==0.6.0
httpx==0.27.0

# Database
sqlite3  


pytest==8.0.0
black==24.1.1
flake8==7.0.0